Remove navigation to home screen from loign handler in AccountLogin screen as setting user object will automatically render home screen for user. Also moved Account layout component to layout folder
